pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mnist-classification"
version = "0.1.0"
description = "Simple MNIST Classification with PyTorch"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "mlflow>=2.8.0",
    "pyyaml>=6.0.0",
]

[tool.setuptools]
packages = ["src"] 